Address PR feedback: move `vectorMatchingGroupKeyFunc` to its own file.
